Move independent components into the components folder
