+++ app/public/components/table/pager.js
@@ -12,11 +12,11 @@
  * or submit itself to any jurisdiction.
  */
 
-import viewButton from '../../../components/common/viewButton.js';
-import { replaceUrlParams } from '../../../utils/utils.js';
+import viewButton from '../common/viewButton.js';
+import { replaceUrlParams } from '../../utils/utils.js';
 import { h, iconChevronBottom } from '/js/src/index.js';
-import itemsCounter from './table/items-counter.js';
-import { RCT } from '../../../config.js';
+import itemsCounter from '../../views/userView/data/table/items-counter.js';
+import { RCT } from '../../config.js';
 
 const { site } = RCT.dataReqParams;
